# docs/swagger.yaml
basePath: /
definitions:
  models.Estate:
    properties:
      id:
        description: Unique identifier for the estate
        type: string
      length:
        description: Length of the estate in 10m plots
        type: integer
      width:
        description: Width of the estate in 10m plots
        type: integer
    type: object
  models.Tree:
    properties:
      estate_id:
        description: ID of the estate this tree belongs to
        type: string
      height:
        description: Height of the tree in meters (1 to 30)
        type: integer
      id:
        description: Unique identifier for the tree
        type: string
      x:
        description: X coordinate of the tree in its plot
        type: integer
      "y":
        description: Y coordinate of the tree in its plot
        type: integer
    type: object
host: localhost:8080
info:
  contact: {}
  description: This is the API documentation for SawitPro Recruitment project.
  title: SawitPro Recruitment API
  version: "1.0"
paths:
  /estate:
    post:
      consumes:
      - application/json
      description: Create a new estate
      parameters:
      - description: Estate
        in: body
        name: estate
        required: true
        schema:
          $ref: '#/definitions/models.Estate'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Estate'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new estate
      tags:
      - estates
  /estate/{id}/drone-plan:
    get:
      description: Calculate the drone's total travel distance with an optional max_distance
        parameter
      parameters:
      - description: Estate ID
        in: path
        name: id
        required: true
        type: string
      - description: Maximum distance the drone can travel
        in: query
        name: max_distance
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Calculate the drone's total travel distance with an optional max_distance
        parameter
      tags:
      - drones
  /estate/{id}/stats:
    get:
      description: Get stats of trees in an estate
      parameters:
      - description: Estate ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: integer
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get stats of trees in an estate
      tags:
      - estates
  /estate/{id}/tree:
    post:
      consumes:
      - application/json
      description: Add a tree to an estate
      parameters:
      - description: Estate ID
        in: path
        name: id
        required: true
        type: string
      - description: Tree
        in: body
        name: tree
        required: true
        schema:
          $ref: '#/definitions/models.Tree'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Tree'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Add a tree to an estate
      tags:
      - trees
swagger: "2.0"